+++ packages/image/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @granite-js/image
 
+## 0.1.34-next.8
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.7
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.6
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.5
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.4
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.3
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.2
+
+### Patch Changes
+
+- fix
+
+## 0.1.34-next.1
+
+### Patch Changes
+
+- fix
+
 ## 0.1.34-next.0
 
 ### Patch Changes